+++ app/controllers/concerns/whitelist.rb
@@ -6,7 +6,7 @@ module Whitelist
     # if Rails.env.development?
     #   ApplicationEnumeration.data_models.inject({}){|hsh, k| hsh.merge!(k.name => k)}.fetch(klass_name)
     # elsif Rails.env.production?
-    ::DATA_MODELS.fetch(klass_name)
+    (::DATA_MODELS.merge('DwcOccurrence' => DwcOccurrence)).fetch(klass_name)
     # else
     #   raise TaxonWorks::Error, 'whitelist attempted in unknown environment'
     # end