+++ app/javascript/vue/tasks/digitize/store/actions/cloneCollectingEvent.js
@@ -8,20 +8,20 @@ import { ActionNames } from './actions'
 import makeCollectingEvent from '@/factory/CollectingEvent.js'
 import makeIdentifier from '@/factory/Identifier.js'
 
-export default ({ dispatch, commit }, collectingEventId) => {
-  CollectingEvent.clone(collectingEventId, { extend: ['roles'] }).then(
-    ({ body }) => {
-      const clonedCE = Object.assign(makeCollectingEvent(), body)
+export default ({ state, dispatch, commit }, params) => {
+  const ceId = state.collecting_event.id
 
-      commit(MutationNames.SetCollectingEvent, clonedCE)
-      commit(
-        MutationNames.SetCollectingEventIdentifier,
-        clonedCE?.identifiers[0] ||
-          makeIdentifier(IDENTIFIER_LOCAL_TRIP_CODE, COLLECTING_EVENT)
-      )
-      commit(MutationNames.SetGeoreferences, [])
-      dispatch(ActionNames.LoadGeoreferences, clonedCE.id)
-      dispatch(ActionNames.SaveDigitalization)
-    }
-  )
+  CollectingEvent.clone(ceId, { ...params }).then(({ body }) => {
+    const clonedCE = Object.assign(makeCollectingEvent(), body)
+
+    commit(MutationNames.SetCollectingEvent, clonedCE)
+    commit(
+      MutationNames.SetCollectingEventIdentifier,
+      clonedCE?.identifiers[0] ||
+        makeIdentifier(IDENTIFIER_LOCAL_TRIP_CODE, COLLECTING_EVENT)
+    )
+    commit(MutationNames.SetGeoreferences, [])
+    dispatch(ActionNames.LoadGeoreferences, clonedCE.id)
+    dispatch(ActionNames.SaveDigitalization)
+  })
 }